✨ コンポーネント間でのstateのリフトアップを実装しました
